+++ src/lib/cc/data.h
@@ -1,4 +1,4 @@
-// Copyright (C) 2010-2024 Internet Systems Consortium, Inc. ("ISC")
+// Copyright (C) 2010-2025 Internet Systems Consortium, Inc. ("ISC")
 //
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -136,7 +136,7 @@ public:
     ///
     /// any is a special type used in list specifications, specifying that the
     /// elements can be of any type.
-    enum types : int {
+    enum types : uint16_t {
         integer = 0,
         real = 1,
         boolean = 2,
